Add avro input format for generic objects. Add end to end test for avro with external jar and external test data.
